Refactor to allow for cwd-based path-to-conf parsing
As initially/intentionally designed, Config::App looks around for a
configuration file from the location of the executable called. This is
desireable because we can thus call tools of a given project outside the
project's directory tree and Config::App will load the configuration of
that project.

However, there are cases (see MojoX::Linda) where we will be calling,
from within a project's tree, an executable that's outside the project's
tree. So in these cases, we should start at cwd.

To make this happen without a bunch of ugly hacks, I had to refactor the
code a bit.